Se mejora obtención de ruts con tope, ya que están aumentando progresivamente - mejora ya que el tipo de dato es biginteger y no long

+++ src/main/java/workcenter/dao/LiquidacionDao.java
@@ -209,7 +209,7 @@ public class LiquidacionDao {
         }
     }
 
-    public List<Long> obtenerGratificadosConTope() {
+    public List<BigInteger> obtenerGratificadosConTope() {
         String nsql = "SELECT rut FROM gratificacion_contope";
         Query q = em.createNativeQuery(nsql);
         return q.getResultList();
